requirements.txt
numpy
pydicom
pillow
pyserde
matplotlib
albumentations 